requirements.txt
@"
Django==4.2.0
djangorestframework==3.14.0
djoser==2.1.0
django-cors-headers==3.14.0
drf-yasg==1.21.5
PyYAML==6.0
Pillow==9.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
"@ | Out-File -FilePath requirements.txt -Encoding UTF8